Fix #21608: Subscription button has no functionality while logged out
From a logged-out user's perspective, the "Subscribe" button
in the comment section of an article page has no functionality.
Although it is visible, it is not clickable.

Expected behavior: Trigger the showLoginModal pop-up, just
like other login-only buttons do.
Refs #21608
