+++ app/javascript/packs/articlePage.jsx
@@ -3,6 +3,9 @@ import { initializeDropdown } from '../utilities/dropdownUtils';
 import { setupBillboardInteractivity } from '../utilities/billboardInteractivity';
 import { embedGists } from '../utilities/gist';
 import { isNativeAndroid, copyToClipboard } from '@utilities/runtime';
+import { h, render } from 'preact';
+import { Snackbar, addSnackbarItem } from '../Snackbar';
+import { initializeUserSubscriptionLiquidTagContent } from '../liquidTags/userSubscriptionLiquidTag';
 
 // Open in new tab backfill
 // We added this behavior on rendering, so this is a backfill for the existing articles
@@ -136,6 +139,72 @@ document
 
 const targetNode = document.querySelector('#comments');
 targetNode && embedGists(targetNode);
+
+
+const { user = null, userStatus } = document.body.dataset;
+const isLoggedIn = userStatus === 'logged-in';
+
+if(isLoggedIn){
+// The Snackbar for the article page, when loggedIn
+  const snackZone = document.getElementById('snack-zone');
+  if (snackZone) {
+    render(<Snackbar lifespan={3} />, snackZone);
+  }
+
+  // eslint-disable-next-line no-restricted-globals
+  top.addSnackbarItem = addSnackbarItem;
+
+  initializeUserSubscriptionLiquidTagContent();
+}
+
+getCsrfToken().then(async () => {
+  const root = document.getElementById('comment-subscription');
+
+  if (!root) {
+    return;
+  }
+  try {
+    const {
+      getCommentSubscriptionStatus,
+      setCommentSubscriptionStatus,
+      CommentSubscription,
+    } = await import('../CommentSubscription');
+
+    const { articleId } = document.getElementById('article-body').dataset;
+
+    let subscriptionType = 'not_subscribed';
+
+    if (isLoggedIn && user !== null) {
+      ({ config: subscriptionType } = await getCommentSubscriptionStatus(
+        articleId,
+      ));
+    }
+
+    const subscriptionRequestHandler = async (type) => {
+      try {
+        const message = await setCommentSubscriptionStatus(articleId, type);
+        addSnackbarItem({ message, addCloseButton: true });
+      } catch (error) {
+        addSnackbarItem({ message: 'Error subscribing to comments.', addCloseButton: true });
+        console.error('Subscription error:', error);
+      }
+    };
+    render(
+      <CommentSubscription
+        subscriptionType={subscriptionType}
+        positionType="static"
+        onSubscribe={subscriptionRequestHandler}
+        onUnsubscribe={subscriptionRequestHandler}
+        isLoggedIn={isLoggedIn}
+      />,
+      root,
+    );
+  } catch (e) {
+    root.innerHTML =
+      '<p className="color-accent-danger">Unable to load Comment Subscription component.<br />Try refreshing the page.</p>';
+  }
+});
+
 
 setupBillboardInteractivity();
 focusOnComments();
